requirements.txt
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
google-auth==2.32.0
